requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.1
attrs==23.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==2.0.12
colorama==0.4.6
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
pytest-timeout==2.1.0
python-dotenv==0.19.0
python-telegram-bot==13.7
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.2
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
zipp==3.15.0
